requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
pyyaml>=6.0
joblib>=1.3
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.6
matplotlib>=3.8
seaborn>=0.13
pytest>=8.0
python-multipart>=0.0.9
